+++ src/lib.rs
@@ -260,6 +260,7 @@ pub mod api;
 pub mod auth;
 pub mod aws;
 pub mod client;
+pub mod cubbyhole;
 pub mod database;
 pub mod error;
 pub mod identity;