+++ vaultrs-tests/tests/api_tests/cubbyhole.rs
@@ -1,25 +1,25 @@
 use crate::common::Test;
 use std::collections::HashMap;
-use vaultrs::{api::kv1::responses::GetSecretResponse, error::ClientError, kv1, sys::mount};
+use vaultrs::{api::cubbyhole::responses::GetSecretResponse, cubbyhole, error::ClientError};
 
 #[tokio::test]
-async fn test_kv1() {
+async fn test_cubbyhole() {
     let test = Test::builder().await;
     let client = test.client();
 
-    // Mount KV v1 secret engine
-    let mount = "kv_v1";
+    // Use pre-mounted cubbyhole secret engine
+    let mount = "cubbyhole";
     let secret_path = "mysecret/foo";
-    mount::enable(client, mount, "kv", None).await.unwrap();
 
     // Create test secrets
     let expected_secret = HashMap::from([("key1", "value1"), ("key2", "value2")]);
-    kv1::set(client, mount, secret_path, &expected_secret)
+    cubbyhole::set(client, mount, secret_path, &expected_secret)
         .await
         .unwrap();
 
     // Read it
-    let read_secret: HashMap<String, String> = kv1::get(client, mount, secret_path).await.unwrap();
+    let read_secret: HashMap<String, String> =
+        cubbyhole::get(client, mount, secret_path).await.unwrap();
 
     println!("{:?}", read_secret);
 
@@ -27,8 +27,9 @@ async fn test_kv1() {
     assert_eq!(read_secret["key2"], expected_secret["key2"]);
 
     // Read it as raw value
-    let read_secret_raw: GetSecretResponse =
-        kv1::get_raw(client, mount, secret_path).await.unwrap();
+    let read_secret_raw: GetSecretResponse = cubbyhole::get_raw(client, mount, secret_path)
+        .await
+        .unwrap();
 
     println!("{:?}", read_secret_raw);
 
@@ -36,16 +37,16 @@ async fn test_kv1() {
     assert_eq!(read_secret_raw.data["key2"], expected_secret["key2"]);
 
     // List secret keys
-    let list_secret = kv1::list(client, mount, "mysecret").await.unwrap();
+    let list_secret = cubbyhole::list(client, mount, "mysecret").await.unwrap();
 
     println!("{:?}", list_secret);
 
     assert_eq!(list_secret.data.keys, vec!["foo"]);
 
     // Delete secret and read again and expect 404 to check deletion
-    kv1::delete(client, mount, secret_path).await.unwrap();
+    cubbyhole::delete(client, mount, secret_path).await.unwrap();
 
-    let r = kv1::get_raw(client, mount, secret_path).await;
+    let r = cubbyhole::get_raw(client, mount, secret_path).await;
 
     match r.expect_err(&format!(
         "Expected error when reading {} after delete.",
@@ -61,18 +62,20 @@ async fn test_kv1() {
 
     let my_secrets = HashMap::from([("key1", "value1"), ("key2", "value2")]);
 
-    kv1::set(client, mount, "my/secrets", &my_secrets)
+    cubbyhole::set(client, mount, "my/secrets", &my_secrets)
         .await
         .unwrap();
 
     let read_secrets: HashMap<String, String> =
-        kv1::get(client, mount, "my/secrets").await.unwrap();
+        cubbyhole::get(client, mount, "my/secrets").await.unwrap();
 
     println!("{:}", read_secrets["key1"]); // value1
 
-    let list_secret = kv1::list(client, mount, "my").await.unwrap();
+    let list_secret = cubbyhole::list(client, mount, "my").await.unwrap();
 
     println!("{:?}", list_secret.data.keys); // [ "secrets" ]
 
-    kv1::delete(client, mount, "my/secrets").await.unwrap();
+    cubbyhole::delete(client, mount, "my/secrets")
+        .await
+        .unwrap();
 }